Fix loading from GCP

+++ dgmr/dgmr.py
@@ -114,7 +114,8 @@ class DGMR(pl.LightningModule, NowcastingModelHubMixin):
 
     def training_step(self, batch, batch_idx):
         images, future_images = batch
-
+        images = images.half()
+        future_images = future_images.half()
         self.global_iteration += 1
         g_opt, d_opt = self.optimizers()
         ##########################
